+++ extensions/omni_httpd/http_worker.c
@@ -242,10 +242,16 @@ void http_worker(Datum db_oid) {
     http_outcome_oid();
 
     {
-      // omni_httpd.handler(int,http_request)
-      List *handler_func = list_make2(makeString("omni_httpd"), makeString("handler"));
-      handler_oid = LookupFuncName(handler_func, -1, (Oid[2]){INT4OID, http_request_oid()}, false);
-      list_free(handler_func);
+      // omni_httpd.handler(int,http_request,http_outcome)
+      List *handler_proc = list_make2(makeString("omni_httpd"), makeString("handler"));
+      List *handler_inputs = list_make2(makeNode(TypeName), makeNode(TypeName));
+      list_nth_node(TypeName, handler_inputs, 0)->typeOid = INT4OID;
+      list_nth_node(TypeName, handler_inputs, 1)->typeOid = http_request_oid();
+      ObjectWithArgs args = {.objname = handler_proc, .objargs = handler_inputs};
+      handler_oid = LookupFuncWithArgs(OBJECT_PROCEDURE, &args, false);
+      Assert(OidIsValid(handler_oid));
+      list_free(handler_proc);
+      list_free(handler_inputs);
     }
 
     {
@@ -898,8 +904,9 @@ static int handler(handler_message_t *msg) {
         fcinfo->args[1].value = HeapTupleGetDatum(request_tuple);
         fcinfo->args[1].isnull = false;
         fcinfo->context = (fmNodePtr)non_atomic_call_context;
-        outcome = FunctionCallInvoke(fcinfo);
-        isnull = fcinfo->isnull;
+        Datum record = FunctionCallInvoke(fcinfo);
+        HeapTupleHeader th = DatumGetHeapTupleHeader(record);
+        outcome = GetAttributeByNum(th, 1, &isnull);
       }
 
       PopActiveSnapshot();